Crate aabb_quadtree

A simple spacial partitioning data structure that allows fast queries for 2-dimensional objects.

As the name implies, the tree is a mapping from axis-aligned-bounding-box => object.

Structs

ItemId
An ID unique to a single QuadTree. This is the object that is returned from queries, and can be used to access the elements stored in the quad tree.

QuadTree
The main QuadTree structure. Mainly supports inserting, removing, and querying objects in 3d space.

Traits

Spatial
An object that has a bounding box.

Type Definitions

QueryResult
Used to determine if a query should keep going or not